Imagenes alpha en Android

+++ Practica1/Engine/src/main/java/es/fdi/ucm/gdv/vdism/maranwi/engine/Graphics.java
@@ -9,7 +9,7 @@ public interface Graphics {
     public void translate(double x,double y);
     public void scale(double x,double y);
     public void drawImage(Image img, int x,int y, int width, int height);
-    public void drawImage(Image img, int x,int y, int width, int height, float alpha);
+    public void drawImage(Image img, int x,int y, int width, int height, int alpha);
     public void setColor(int rgba);
     public void setColor(int r,int g, int b,int a);
     public void setColor(Color color);
